==> 25_assign_region.py <==
# Auto-generated from 25_assign_region.ipynb
# Do not edit this file directly; edit the notebook instead.

# %%

import pandas as pd

# %%

df = pd.read_csv('df_check_4.csv')
df

# %%

# UN M.49に基づく世界地理区分（日本語表記に準じた例）
import numpy as np
data = [
    {"国名": "ケニア", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ベトナム", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "モンゴル", "地域": "アジア", "地域詳細": "東アジア"},
    {"国名": "中華人民共和国", "地域": "アジア", "地域詳細": "東アジア"},
    {"国名": "ガーナ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ニカラグア", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "カンボジア", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "ネパール", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "タンザニア", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "マリ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "インドネシア", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "エジプト", "地域": "アフリカ", "地域詳細": "北アフリカ"},
    {"国名": "ボスニア・ヘルツェゴビナ", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "インド", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "ザンビア", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "ナイジェリア", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "トンガ", "地域": "大洋州", "地域詳細": "ポリネシア"},
    {"国名": "パラオ", "地域": "大洋州", "地域詳細": "ミクロネシア"},
    {"国名": "カメルーン", "地域": "アフリカ", "地域詳細": "中央アフリカ"},
    {"国名": "グアテマラ", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "サモア", "地域": "大洋州", "地域詳細": "ポリネシア"},
    {"国名": "モザンビーク", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ボリビア", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "バングラデシュ", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "セネガル", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ブルキナファソ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ウガンダ", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "パラグアイ", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "パキスタン", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "マレーシア", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "ラオス", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "モロッコ", "地域": "アフリカ", "地域詳細": "北アフリカ"},
    {"国名": "アンゴラ", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "ホンジュラス", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "イラン", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "カザフスタン", "地域": "アジア", "地域詳細": "中央アジア"},
    {"国名": "モーリシャス", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ペルー", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "フィリピン", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "スリランカ", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "チュニジア", "地域": "アフリカ", "地域詳細": "北アフリカ"},
    {"国名": "ガボン", "地域": "アフリカ", "地域詳細": "中央アフリカ"},
    {"国名": "東ティモール", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "ニジェール", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ベナン", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "南アフリカ共和国", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "エリトリア", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ガンビア", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "キルギス", "地域": "アジア", "地域詳細": "中央アジア"},
    {"国名": "ブータン", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "アフガニスタン", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "ヨルダン", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "エチオピア", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "キリバス", "地域": "大洋州", "地域詳細": "ミクロネシア"},
    {"国名": "ツバル", "地域": "大洋州", "地域詳細": "ポリネシア"},
    {"国名": "マラウイ", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "ルワンダ", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "エルサルバドル", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "タイ", "地域": "アジア", "地域詳細": "東南アジア"},
    {"国名": "エクアドル", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "パプアニューギニア", "地域": "大洋州", "地域詳細": "メラネシア"},
    {"国名": "コスタリカ", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "アルゼンチン", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "ブラジル", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "チリ", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "パレスチナ", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "サウジアラビア", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "トルコ", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "アゼルバイジャン", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "シエラレオネ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ポーランド", "地域": "欧州", "地域詳細": "東欧"},
    {"国名": "ソロモン", "地域": "大洋州", "地域詳細": "メラネシア"},
    {"国名": "セルビア", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "ミクロネシア連邦", "地域": "大洋州", "地域詳細": "ミクロネシア"},
    {"国名": "ウズベキスタン", "地域": "アジア", "地域詳細": "中央アジア"},
    {"国名": "モンテネグロ", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "バヌアツ", "地域": "大洋州", "地域詳細": "メラネシア"},
    {"国名": "モルディブ", "地域": "アジア", "地域詳細": "南アジア"},
    {"国名": "セントビンセント", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "モーリタニア", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ウクライナ", "地域": "欧州", "地域詳細": "東欧"},
    {"国名": "モルドバ", "地域": "欧州", "地域詳細": "東欧"},
    {"国名": "ジョージア", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "スロバキア", "地域": "欧州", "地域詳細": "東欧"},
    {"国名": "マダガスカル", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "北マケドニア共和国", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "ギニア", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "セントクリストファー・ネービス", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "ミャンマー", "地域": "アジア", "地域詳細": "東南アジア"},

    # 以下、見出し行等は除外
    {"国名": "ドミニカ共和国", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "メキシコ", "地域": "中南米", "地域詳細": "北中米"},
    {"国名": "パナマ", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "アルジェリア", "地域": "アフリカ", "地域詳細": "北アフリカ"},
    {"国名": "ブルガリア", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "スリナム", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "カーボベルデ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ジブチ", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "セーシェル", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "セントルシア", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "タジキスタン", "地域": "アジア", "地域詳細": "中央アジア"},
    {"国名": "アルメニア", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "ルーマニア", "地域": "欧州", "地域詳細": "東欧"},
    {"国名": "ジャマイカ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "レソト", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "ブルンジ", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ガイアナ", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "マーシャル", "地域": "大洋州", "地域詳細": "ミクロネシア"},
    {"国名": "ドミニカ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "アンティグア・バーブーダ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "フィジー", "地域": "大洋州", "地域詳細": "メラネシア"},
    {"国名": "アルバニア", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "グレナダ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "レバノン", "地域": "アジア", "地域詳細": "西アジア"},

    # "(アジア)"、"(アフリカ)"等の見出し行は除外
    {"国名": "コンゴ民主共和国", "地域": "アフリカ", "地域詳細": "中央アフリカ"},
    {"国名": "エスワティニ", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "ナミビア", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "バルバドス", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "コロンビア", "地域": "中南米", "地域詳細": "南アメリカ"},
    {"国名": "キューバ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "スーダン", "地域": "アフリカ", "地域詳細": "北アフリカ"},
    {"国名": "ベリーズ", "地域": "中南米", "地域詳細": "中央アメリカ"},
    {"国名": "ギニアビサウ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ハイチ", "地域": "中南米", "地域詳細": "カリブ海"},
    {"国名": "コソボ", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "トーゴ", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "コモロ", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "ジンバブエ", "地域": "アフリカ", "地域詳細": "南部アフリカ"},
    {"国名": "クロアチア", "地域": "欧州", "地域詳細": "南ヨーロッパ"},
    {"国名": "イラク", "地域": "アジア", "地域詳細": "西アジア"},
    {"国名": "南スーダン", "地域": "アフリカ", "地域詳細": "東アフリカ"},
    {"国名": "コートジボワール", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "リベリア", "地域": "アフリカ", "地域詳細": "西アフリカ"},
    {"国名": "ボツワナ", "地域": "アフリカ", "地域詳細": "南部アフリカ"},

    # 複数地域
    {"国名": "(大洋州)", "地域": "大洋州", "地域詳細": np.NaN},
    {"国名": "(中南米)", "地域": "中南米", "地域詳細": np.NaN},
    {"国名": "(アジア)", "地域": "アジア", "地域詳細": np.NaN},
    {"国名": "(アフリカ)", "地域": "アフリカ", "地域詳細": np.NaN},
    {"国名": "アフリカ開発銀行", "地域": "アフリカ", "地域詳細": np.NaN},
    {"国名": "(中東)", "地域": "アジア", "地域詳細": np.NaN},
]

df_country = pd.DataFrame(data)
df_country

# %%

df = pd.merge(df, df_country, on='国名', how='left')

# %%

# 地域がNULLでないことを確認 
assert len(df[df['地域'].isna()]) == 0

# %%

# 複数国あるため、地域詳細はNULL許容
df[df['地域詳細'].isna()]

# %%

df.head(10)

# %%

df.to_csv('df_check_5.csv')
print(f"レコード数: {len(df)}")
