==> 55_assign_ex_eval_flg.py <==
# Auto-generated from 55_assign_ex_eval_flg.ipynb
# Do not edit this file directly; edit the notebook instead.

# %%

import pandas as pd
df = pd.read_csv("df_check_10.csv", index_col=0)
df

# %%

# 評価会社の全量確認
li = df['評価会社'].unique().tolist()
li

# %%

# targetsにあるものは内部評価としてカウント
targets = ['部', '支所', '事務所','国際協力機構','課', 'JICA']

result = [
    x for x in li 
    if isinstance(x, str) and any(t in x for t in targets)
]
result

# %%

# 外部評価だけであることを確認
others = [x for x in li if x not in result]
others

# %%

# 外部評価フラグを設定
df['external_eval_flg'] = df['評価会社'].isin(others)
col = ['評価会社', 'external_eval_flg']
df[col]

# %%

df.to_csv('df_check_11.csv')
print(f"レコード数: {len(df)}")
# %%

pd.set_option('display.max_rows', 900)

# %%

cols = ['評価会社','external_eval_flg']
df[cols].drop_duplicates().sort_values(['external_eval_flg'])

# %%
