==> 70_assign_rate.py <==
# Auto-generated from 99_rename_df.ipynb
# Do not edit this file directly; edit the notebook instead.

# %%
# 実績と計画値の割合の特徴量を付与する
import pandas as pd
df = pd.read_csv("df_check_13.csv", index_col=0)
df.head(5)
# %%

df['cost_rate'] = df['プロジェクトコスト_実績_int']/df['プロジェクトコスト_計画時_int']
df['duration_rate'] = df['実績_プロジェクト期間（月）']/df['事前評価時_プロジェクト期間（月）']
# %%
df.to_csv('df_check_14.csv')
