==> 99_rename_df.py <==
# Auto-generated from 99_rename_df.ipynb
# Do not edit this file directly; edit the notebook instead.

# %%

import pandas as pd
df = pd.read_csv("df_check_14.csv", index_col=0)


# %%

# 総合評価を数値に変換
df['total_eval'] = df['総合評価'].map({'非常に高い': 4, '高い': 3, '中程度': 2,'一部課題がある': 2, '一部に課題があると判断される': 2, '低い': 1,'非常に低い': 1})


# %%

rename = {
    '評価年度':'eval_year',
    '評価会社':'evaluator_group',
    '評価者':'evaluator',
    'プロジェクトコスト_計画時_int':'project_cost_plan',
    'プロジェクトコスト_実績_int':'project_cost_act',
    '事前評価時_プロジェクト期間（月）':'project_duration_plan', 
    '実績_プロジェクト期間（月）':'project_duration_act', 
    '地域詳細':'region_detail',
    '地域':'region',
    '事業形態':'type',
}
df=df.rename(columns=rename)
df

# %%

df['total_eval'].unique()

# %%

df.to_csv('df_check_99.csv')
print(f"レコード数: {len(df)}")
